+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v0.1.6] - 2026-03-04
+- Improvement: Search bar filter dialog implemented.
+
 ## [v0.1.5] - 2026-03-04
 - Bugfix: Fixed Google OAuth login.
 - Improvement: Encapsulated Google OAuth login.